# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Showings Route Mapper
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PropertyAddress {
  /** Full address string */
  address: string;
  /** Optional custom label (e.g. "Starting Point", "123 Oak St") */
  label?: string;
}

export interface OptimizeRouteBody {
  /**
   * List of property addresses to visit
   * @minItems 2
   */
  properties: PropertyAddress[];
  /** Optional starting address for the agent (defaults to first property) */
  startAddress?: string;
}

export interface GeocodedLocation {
  address: string;
  label?: string;
  lat: number;
  lng: number;
  /** Full formatted address returned by geocoder */
  displayName: string;
}

export interface RouteLeg {
  from: GeocodedLocation;
  to: GeocodedLocation;
  distanceKm: number;
  durationMinutes: number;
  /** 1-based order of this leg in the route */
  order: number;
}

export interface OptimizedRoute {
  /** Ordered list of stops to visit */
  stops: GeocodedLocation[];
  legs: RouteLeg[];
  totalDistanceKm: number;
  totalDurationMinutes: number;
  /** Array of [lat, lng] pairs forming the full route path for map rendering */
  polyline: number[][];
}

export interface ShareRouteBody {
  route: OptimizedRoute;
  /** Optional start time in HH:MM format */
  startTime?: string;
}

export interface ShareRouteResponse {
  shareId: string;
  shareUrl: string;
}

export interface GeocodeBody {
  address: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}