# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Showings Route Mapper
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Accepts a list of property addresses, geocodes them, computes optimized visit order, and returns route with distances and travel times
 * @summary Optimize route for property showings
 */
export const optimizeRouteBodyPropertiesMin = 2;

export const OptimizeRouteBody = zod.object({
  properties: zod
    .array(
      zod.object({
        address: zod.string().describe("Full address string"),
        label: zod
          .string()
          .optional()
          .describe(
            'Optional custom label (e.g. \"Starting Point\", \"123 Oak St\")',
          ),
      }),
    )
    .min(optimizeRouteBodyPropertiesMin)
    .describe("List of property addresses to visit"),
  startAddress: zod
    .string()
    .optional()
    .describe(
      "Optional starting address for the agent (defaults to first property)",
    ),
});

export const optimizeRouteResponsePolylineItemMin = 2;
export const optimizeRouteResponsePolylineItemMax = 2;

export const OptimizeRouteResponse = zod.object({
  stops: zod
    .array(
      zod.object({
        address: zod.string(),
        label: zod.string().optional(),
        lat: zod.number(),
        lng: zod.number(),
        displayName: zod
          .string()
          .describe("Full formatted address returned by geocoder"),
      }),
    )
    .describe("Ordered list of stops to visit"),
  legs: zod.array(
    zod.object({
      from: zod.object({
        address: zod.string(),
        label: zod.string().optional(),
        lat: zod.number(),
        lng: zod.number(),
        displayName: zod
          .string()
          .describe("Full formatted address returned by geocoder"),
      }),
      to: zod.object({
        address: zod.string(),
        label: zod.string().optional(),
        lat: zod.number(),
        lng: zod.number(),
        displayName: zod
          .string()
          .describe("Full formatted address returned by geocoder"),
      }),
      distanceKm: zod.number(),
      durationMinutes: zod.number(),
      order: zod.number().describe("1-based order of this leg in the route"),
    }),
  ),
  totalDistanceKm: zod.number(),
  totalDurationMinutes: zod.number(),
  polyline: zod
    .array(
      zod
        .array(zod.number())
        .min(optimizeRouteResponsePolylineItemMin)
        .max(optimizeRouteResponsePolylineItemMax),
    )
    .describe(
      "Array of [lat, lng] pairs forming the full route path for map rendering",
    ),
});

/**
 * Stores an optimized route and returns a share ID and URL
 * @summary Save a route for sharing
 */
export const shareRouteBodyRoutePolylineItemMin = 2;
export const shareRouteBodyRoutePolylineItemMax = 2;

export const ShareRouteBody = zod.object({
  route: zod.object({
    stops: zod
      .array(
        zod.object({
          address: zod.string(),
          label: zod.string().optional(),
          lat: zod.number(),
          lng: zod.number(),
          displayName: zod
            .string()
            .describe("Full formatted address returned by geocoder"),
        }),
      )
      .describe("Ordered list of stops to visit"),
    legs: zod.array(
      zod.object({
        from: zod.object({
          address: zod.string(),
          label: zod.string().optional(),
          lat: zod.number(),
          lng: zod.number(),
          displayName: zod
            .string()
            .describe("Full formatted address returned by geocoder"),
        }),
        to: zod.object({
          address: zod.string(),
          label: zod.string().optional(),
          lat: zod.number(),
          lng: zod.number(),
          displayName: zod
            .string()
            .describe("Full formatted address returned by geocoder"),
        }),
        distanceKm: zod.number(),
        durationMinutes: zod.number(),
        order: zod.number().describe("1-based order of this leg in the route"),
      }),
    ),
    totalDistanceKm: zod.number(),
    totalDurationMinutes: zod.number(),
    polyline: zod
      .array(
        zod
          .array(zod.number())
          .min(shareRouteBodyRoutePolylineItemMin)
          .max(shareRouteBodyRoutePolylineItemMax),
      )
      .describe(
        "Array of [lat, lng] pairs forming the full route path for map rendering",
      ),
  }),
  startTime: zod
    .string()
    .optional()
    .describe("Optional start time in HH:MM format"),
});

export const ShareRouteResponse = zod.object({
  shareId: zod.string(),
  shareUrl: zod.string(),
});

/**
 * Returns the stored optimized route for a given share ID
 * @summary Retrieve a shared route
 */
export const GetSharedRouteParams = zod.object({
  shareId: zod.coerce.string(),
});

export const getSharedRouteResponsePolylineItemMin = 2;
export const getSharedRouteResponsePolylineItemMax = 2;

export const GetSharedRouteResponse = zod.object({
  stops: zod
    .array(
      zod.object({
        address: zod.string(),
        label: zod.string().optional(),
        lat: zod.number(),
        lng: zod.number(),
        displayName: zod
          .string()
          .describe("Full formatted address returned by geocoder"),
      }),
    )
    .describe("Ordered list of stops to visit"),
  legs: zod.array(
    zod.object({
      from: zod.object({
        address: zod.string(),
        label: zod.string().optional(),
        lat: zod.number(),
        lng: zod.number(),
        displayName: zod
          .string()
          .describe("Full formatted address returned by geocoder"),
      }),
      to: zod.object({
        address: zod.string(),
        label: zod.string().optional(),
        lat: zod.number(),
        lng: zod.number(),
        displayName: zod
          .string()
          .describe("Full formatted address returned by geocoder"),
      }),
      distanceKm: zod.number(),
      durationMinutes: zod.number(),
      order: zod.number().describe("1-based order of this leg in the route"),
    }),
  ),
  totalDistanceKm: zod.number(),
  totalDurationMinutes: zod.number(),
  polyline: zod
    .array(
      zod
        .array(zod.number())
        .min(getSharedRouteResponsePolylineItemMin)
        .max(getSharedRouteResponsePolylineItemMax),
    )
    .describe(
      "Array of [lat, lng] pairs forming the full route path for map rendering",
    ),
});

/**
 * Returns lat/lng for a given address string
 * @summary Geocode a single address
 */
export const GeocodeAddressBody = zod.object({
  address: zod.string(),
});

export const GeocodeAddressResponse = zod.object({
  address: zod.string(),
  label: zod.string().optional(),
  lat: zod.number(),
  lng: zod.number(),
  displayName: zod
    .string()
    .describe("Full formatted address returned by geocoder"),
});